RoadBuddy/requirements_qwen.txt
# Qwen2-VL Requirements for RoadBuddy Challenge

# Core deep learning
torch>=2.1.0
torchvision>=0.16.0

# Transformers and Qwen2-VL
transformers>=4.37.0
qwen-vl-utils
accelerate>=0.25.0

# Quantization and optimization
bitsandbytes>=0.41.0
optimum>=1.16.0

# Fine-tuning
peft>=0.7.0
trl>=0.7.0

# Data processing
datasets>=2.16.0
opencv-python>=4.8.0
pandas>=2.0.0
pillow>=10.0.0
numpy>=1.24.0
av>=10.0.0

# Utilities
tqdm>=4.65.0
einops>=0.7.0
safetensors>=0.4.0
pyyaml>=6.0

# Optional: Flash Attention for speed (requires specific CUDA setup)
# flash-attn>=2.0.0

# OCR (optional for traffic sign reading)
easyocr>=1.7.0

